requirements.txt
yfinance
pandas
streamlit
openai
plotly
numpy
google-generativeai
